% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralClustering.R
\name{SpectralClustering}
\alias{SpectralClustering}
\title{Spectral clustering}
\usage{
SpectralClustering(
  Dt,
  S,
  Laplacian = c("sym", "rw"),
  k = 2:10,
  sigma = 1,
  sparse = F,
  OptKmeans = c("SSw", "PseudoF", "DBindex", "DunnIndex"),
  LocalScaling = c("no", "kernlab", "k = 7", "kNN"),
  kNNLocalScaling = NULL,
  ...
)
}
\arguments{
\item{Dt}{Object that contains the data.}

\item{S}{The similarity matrix.}

\item{Laplacian}{Normalization of the Laplacian matrix. Symmetric or random walk.}

\item{k}{Number of clusters}

\item{sigma}{Sigma in the Gaussian similarity measure.}

\item{sparse}{Indicates whether sparse matrices are used.}

\item{OptKmeans}{Measure to select the optimal k-means clustering solution.}

\item{LocalScaling}{Whether local scaling has to be applied.}

\item{kNNLocalScaling}{Number of nearest neighbors when using local scaling.}

\item{...}{Arguments passed to kmeans.}
}
\description{
Implementation of the spectral clustering algorithm of Ng et al. (2001)
}
\examples{
library(kernlab)
data(spirals)
SpectralClusteringGridSearch(spirals, k = 2)

Dt = as.data.frame(T4cluster::genSMILEY())[, -3]
x  = SpectralClustering(Dt, k = 4, LocalScaling = "kernlab")
x
plot(x)
}
\references{
Ng, A., Jordan, M. & Weiss, Y. (2001). On spectral clustering: Analysis and an algorithm. \emph{Advances in Neural Information Processing Systems}, 14.
}
